Reverted pending storage folders removal, updated deletePendingStateCommand to only remove unfinalized states

+++ src/commands/protocols/update/receiver/delete-pending-state-command.js
@@ -1,26 +1,40 @@
 import Command from '../../../command.js';
-import { ERROR_TYPE } from '../../../../constants/constants.js';
+import { ERROR_TYPE, PENDING_STORAGE_REPOSITORIES } from '../../../../constants/constants.js';
 
 class DeletePendingStateCommand extends Command {
     constructor(ctx) {
         super(ctx);
+        this.blockchainModuleManager = ctx.blockchainModuleManager;
         this.pendingStorageService = ctx.pendingStorageService;
 
         this.errorType = ERROR_TYPE.UPDATE.UPDATE_DELETE_PENDING_STATE_ERROR;
     }
 
     async execute(command) {
-        const { repository, blockchain, contract, tokenId, assertionId, operationId } =
-            command.data;
+        const { blockchain, contract, tokenId, assertionId, operationId } = command.data;
 
-        await this.pendingStorageService.removeCachedAssertion(
-            repository,
+        const assetStates = this.blockchainModuleManager.getAssertionIds(
             blockchain,
             contract,
             tokenId,
-            assertionId,
-            operationId,
         );
+
+        if (!assetStates.includes(assertionId)) {
+            for (const repository of [
+                PENDING_STORAGE_REPOSITORIES.PUBLIC,
+                PENDING_STORAGE_REPOSITORIES.PRIVATE,
+            ]) {
+                // eslint-disable-next-line no-await-in-loop
+                await this.pendingStorageService.removeCachedAssertion(
+                    repository,
+                    blockchain,
+                    contract,
+                    tokenId,
+                    assertionId,
+                    operationId,
+                );
+            }
+        }
 
         return Command.empty();
     }
